.env ok

+++ django/friend/views.py
@@ -1,17 +1,27 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
-from .models import Friendship
 from django.contrib import messages
 from django.db.models import Q
-from .models import Player
-
 from django.utils import timezone
-# Create your views here.
+from django.urls import reverse
+from .models import Friendship, Player
+from player.jwt import generate_jwt, decode_jwt, token_user, set_jwt_token
+import logging
 
+
+@login_required
 def index(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     return render(request, "friend/index.html")
 
+@login_required
 def add(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     if request.method == 'POST':
         you = request.user
         you.last_login = timezone.now()
@@ -33,39 +43,63 @@ def add(request):
         # Handle the submitted username as needed
         # For example, you can query the User model or process it in other ways
     return render(request, "friend/add.html")
-    
+
+@login_required    
 def delete_friend(request, id_friendship):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     request = Friendship.objects.get(id=id_friendship)
     request.delete()
     return redirect("friend:list")
 
+@login_required
 def accept(request, id_friendship):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     request = Friendship.objects.get(id=id_friendship)
     request.status = 'accepted'
     request.save()
     return redirect("friend:pending")
 
+@login_required
 def refuse(request, id_friendship):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     request = Friendship.objects.get(id=id_friendship)
     request.status = 'refused'
     request.save()
     return redirect("friend:pending")
 
+@login_required
 def list(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     you = request.user
     friends = Friendship.objects.filter(Q(user=you, status='accepted') | Q(friend=you, status='accepted'))
     you.last_login = timezone.now()
     you.save()
     return render(request, "friend/list.html", {'friends':friends, 'you':you})
 
+@login_required
 def pending(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     you = request.user
     friends = Friendship.objects.filter(Q(friend=you, status='pending'))
     you.last_login = timezone.now()
     you.save()
     return render(request, "friend/pending.html", {'friends':friends, 'you':you})
 
+@login_required
 def refused(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     you = request.user
     friends = Friendship.objects.filter(Q(friend=you, status='refused'))
     you.last_login = timezone.now()
